Adding custom audio button
Implementation of a custom audio player using the code provided at https://code-boxx.com/html-custom-audio-player/. Functionally, it's the same as the HTML audio player, but it is able to be customized more. The next step is to figure out if there's a way to expand upon this to make multiple sound files play with a single click of the play button.